Removed previously declared pages of page_manager.i_am_member_of_spaces and page_manager.i_am_member_of_circles, removed the block LivingSpacesGroupLinksIAmMemberBlock, replaced it with the view's tab menu.
Added the tab menu for the page 'I am a member of Spaces' and made it as default tab for menu item 'All spaces & circles' (internal:/membership), added second tab menu for the page 'I am Member of Circles' (, internal:/membership/circles). Created the display 'menu_member_of_spaces' in the Spaces view to show results from it in the main menu. Created settings for the 'I am Member of Circles' in the rewrite folder of 'living_spaces_circles' module. Created settings for the 'menu_member_of_spaces' in the rewrite folder of 'living_spaces_menus' module.

Added the parent item Spaces to the main menu, exported default content configs for it. Created static menu item 'All spaces & circles' leads to 'internal:/membership' in the main menu, also use it as group for Spaces and Circles membership links.

Added links 'living_spaces_menus.space' with class MembershipLink and Derivative for it to show the links from the 'spaces.menu_member_of_spaces' view as children items in the Spaces section of the main menu. Added the cache dependency of the user to this links. Added the hook 'living_spaces_menus_user_update()' to rebuild the main menu if joined spaces of the user was changed.

Fixed view's filter LivingSpacesGroupInUserJoinedSpaces to use it with anonymous user.